Improve test scripts
